// src/app/interfaces/otretro.ts
export interface OtRetro{
    id?: number;
    folio: number;
    retroexcavadora: string;	
    operador: string;
    num_trabajador: string;
    fecha: string;
    turno: string;
    hora_inicial: string;	
    ubicacion_inicial: string;
    horometro_inicial: string;
    horometro_final: string;
    ubicacion_final: string;
    hora_final: string;
    combustible: string;
    aceite_hidraulico?: string;
    aceite_transmision?: string;
    aceite_motor?: string;
    codigo_operacion1: string;
    nivel1: string;
    lugar1: string;
    empresa1: string;
    hora_inicial1: string;
    hora_final1: string;
    horometro_inicial1?: string;
    horometro_final1?: string;
    breve_descripcion1: string;
    codigo_operacion2: string;
    nivel2: string;
    lugar2: string;
    empresa2: string;
    hora_inicial2: string;
    hora_final2: string;
    horometro_inicial2?: string;
    horometro_final2?: string;
    breve_descripcion2: string;
    codigo_operacion3: string;
    nivel3: string;
    lugar3: string;
    empresa3: string;
    hora_inicial3: string;
    hora_final3: string;
    horometro_inicial3?: string;
    horometro_final3?: string;
    breve_descripcion3: string;
    codigo_operacion4: string;
    nivel4: string;
    lugar4: string;
    empresa4: string;
    hora_inicial4: string;
    hora_final4: string;
    horometro_inicial4?: string;
    horometro_final4?: string;
    breve_descripcion4: string;
    codigo_operacion5: string;
    nivel5: string;
    lugar5: string;
    empresa5: string;
    hora_inicial5: string;
    hora_final5: string;
    horometro_inicial5?: string;
    horometro_final5?: string;
    breve_descripcion5: string;
    codigo_operacion6: string;
    nivel6: string;
    lugar6: string;
    empresa6: string;	
    hora_inicial6: string;	
    hora_final6: string;	
    horometro_inicial6?: string;
    horometro_final6?: string;
    breve_descripcion6: string;	
    codigo_operacion7: string;	
    nivel7: string;	
    lugar7: string;	
    empresa7: string;	
    hora_inicial7: string;	
    hora_final7: string;
    horometro_inicial7?: string;
    horometro_final7?: string;
    breve_descripcion7: string;
    codigo_operacion8: string;
    nivel8: string;
    lugar8: string;
    empresa8: string;	
    hora_inicial8: string;	
    hora_final8: string;
    horometro_inicial8?: string;
    horometro_final8?: string;
    breve_descripcion8: string;	
    nombre_sup_avar?: string;	
    nombre_sup_contratista?: string;	
    permiso_edicion?: string;	
    revisado_valido?: string;	
    created_at?: string;	
    updated_at?: string;
}